fix(adt): batch-42-backend-utils-adt — 15 bug fixes (#2153)
Bug-inventory batch batch-42-backend-utils-adt (utils/adt, audit-codex rows).

- mac8: macaddr8_in parses the raw cstring bytes (no UTF-8 lossy pass), so
  a trailing non-UTF-8 byte after six octets is left unread as in C, and the
  rejected input is echoed byte-for-byte.
- misc: atooid follows strtoul (whitespace, sign, ULONG_MAX saturation,
  32-bit truncation); directory_is_empty reports a readdir failure like
  ReadDir; pg_input_error_info checks the composite result type before
  parsing the type name.
- multirangetypes: multirange_constructor0 rejects arguments (XX000); the
  NULL-member sanity checks in constructor1/2 are elog(ERROR)s (XX000);
  multirange_recv grows the pointer array on demand instead of reserving
  16 bytes per entry (C admits 8 bytes per pointer).
- numeric: accum_sum_rescale frees the replaced digit buffers.
- orderedsetaggs: mode/rank/dense_rank final scans check for interrupts.
- pg_locale: strncoll_libc's combined buffer and icu_to_uchar's UTF-16
  buffer get palloc's MaxAllocSize admission.
- pgstatfuncs: pg_stat_get_replication_slot clips the slot name at 63 bytes
  like namestrcpy, on any byte boundary.
- rangetypes: range_bound_escape reserves the exact escaped length instead
  of 2*len+2, so a 512 MiB bound outputs as in C.
- rangetypes_gist: LIMIT_RATIO is a double as in C, so a float4 ratio of
  exactly 3/10 is accepted.
- regexp: regexp_matches/regexp_split_to_table keep an owned snapshot across
  SRF calls and build one row per call, dropping the nmatches*24 ceiling.

+++ crates/backend/utils/adt/rangetypes_gist/src/lib.rs
@@ -46,7 +46,9 @@ const CLS_CONTAIN_EMPTY: usize = 4;
 const CLS_EMPTY: usize = 8;
 const CLS_COUNT: usize = 9;
 
-const LIMIT_RATIO: f32 = 0.3;
+// C's macro is a double literal: the float4 ratio is promoted before the
+// compare, so a ratio of exactly 3/10 (rounded up as float4) passes.
+const LIMIT_RATIO: f64 = 0.3;
 
 const INFINITE_BOUND_PENALTY: f32 = 2.0;
 const CONTAIN_EMPTY_PENALTY: f32 = 1.0;
@@ -962,7 +964,7 @@ fn consider_split(
     let right_count = context.entries_count - left_count;
 
     let ratio = left_count.min(right_count) as f32 / context.entries_count as f32;
-    if ratio > LIMIT_RATIO {
+    if ratio as f64 > LIMIT_RATIO {
         let overlap = if context.has_subtype_diff {
             call_subtype_diff(mcx, cache, left_upper.val, right_lower.val)? as f32
         } else {
